Enhanced google authentication
